add evaluation and losses

+++ src/squad_f1.py
@@ -41,9 +41,9 @@ def compute_f1(a_gold, a_pred):
     num_same = sum(common.values())
     if len(gold_toks) == 0 or len(pred_toks) == 0:
         # If either is no-answer, then F1 is 1 if they agree, 0 otherwise
-        return int(gold_toks == pred_toks)
+        return float(gold_toks == pred_toks)
     if num_same == 0:
-        return 0
+        return 0.
     precision = 1.0 * num_same / len(pred_toks)
     recall = 1.0 * num_same / len(gold_toks)
     f1 = (2 * precision * recall) / (precision + recall)
